Pt_5_2: Adding new books. Search by author
Using the previous made file, program asks user how many new books they
want to add to the list and asks user to input data for a new books.
After that, program asks user to input an author, whose books they want
to search in the list. If there are no books of user's author, relevant
message will be outputted